Vendedores [sin poder crear ni actualizar]

+++ admin/index.php
@@ -2,24 +2,23 @@
 
 
 require '../includes/app.php';
+require '../classes/vendedor.php';
 estaAutenticado();
 
 use App\Propiedad;
+use App\Vendedor;
+
+// Configurar conexión a la base de datos
+$db = conectarDB();
+
+// Pasar la conexión a la clase vendedor
+vendedor::setDB($db);
 
 //Implementar un método para obtener todas las propiedades
 $propiedades = Propiedad::all();
+$vendedores = Vendedor::all();
 
-// //importar la conexion
-// require '../includes/config/database.php';
-// $db = conectarDB();
 
-// //Escribir el query
-// $query = "SELECT * FROM propiedades";
-
-// //Consultar la base de datos
-// $resultadoConsulta = mysqli_query($db, $query);
-
-//Muestra mensaje condicional
 $resultado = $_GET['resultado'] ?? null;
 
 
@@ -29,33 +28,65 @@ $resultado = $_GET['resultado'] ?? null;
 include '../includes/templates/header.php';
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+
+
     $id = $_POST['id'] ?? null;
 
     if ($id) {
         $id = filter_var($id, FILTER_VALIDATE_INT);
 
         if ($id) {
-            // Obtener la propiedad antes de eliminar
-            $querySelect = "SELECT imagen FROM propiedades WHERE id = $id";
-            $resultadoSelect = mysqli_query($db, $querySelect);
 
-            if ($resultadoSelect && $propiedad = mysqli_fetch_assoc($resultadoSelect)) {
-                $imagen = $propiedad['imagen'];
-                if ($imagen && file_exists("../imagenes/" . $imagen)) {
-                    unlink("../imagenes/" . $imagen);
+            $tipo = $_POST['tipo'];
+
+            if (validarTipoContenido($tipo)) {
+                //compara lo que vamos a eliminar
+                if ($tipo === 'vendedor') {
+                    // Obtener el vendedor antes de eliminar
+                    // $querySelect = "SELECT imagen FROM vendedores WHERE vendedor_id = $id";
+                    // No intentamos obtener 'imagen' porque no existe en la tabla vendedores
+                    $queryDelete = "DELETE FROM vendedores WHERE vendedor_id = $id";
+                    $resultadoConsulta = mysqli_query($db, $queryDelete);
+
+                    if ($resultadoConsulta) {
+                        header('Location: /admin?resultado=4');
+                    } else {
+                        echo "Error al eliminar vendedor " . mysqli_error($db);
+                    }
+                    // Actualizar el vendedor
+                    $queryUpdate = "UPDATE vendedores  WHERE vendedor_id = $id";
+                    $resultadoConsultaUpdate = mysqli_query($db, $queryUpdate);
+                    if ($resultadoConsultaUpdate) {
+                        header('Location: /admin?resultado=5');
+                    } else {
+                        echo "Error al actualizar vendedor " . mysqli_error($db);
+                    }
+
+                } else if ($tipo === 'propiedad') {
+
+                    // Obtener la propiedad antes de eliminar
+                    $querySelect = "SELECT imagen FROM propiedades WHERE id = $id";
+                    $resultadoSelect = mysqli_query($db, $querySelect);
+
+                    if ($resultadoSelect && $propiedad = mysqli_fetch_assoc($resultadoSelect)) {
+                        $imagen = $propiedad['imagen'];
+                        if ($imagen && file_exists("../imagenes/" . $imagen)) {
+                            unlink("../imagenes/" . $imagen);
+                        }
+
+                        // Eliminar la propiedad
+                        $queryDelete = "DELETE FROM propiedades WHERE id = $id";
+                        $resultadoConsulta = mysqli_query($db, $queryDelete);
+
+                        if ($resultadoConsulta) {
+                            header('Location: /admin?resultado=3');
+                        } else {
+                            echo "Error al eliminar la propiedad: " . mysqli_error($db);
+                        }
+                    } else {
+                        echo "No se encontró la propiedad.";
+                    }
                 }
-
-                // Eliminar la propiedad
-                $queryDelete = "DELETE FROM propiedades WHERE id = $id";
-                $resultadoConsulta = mysqli_query($db, $queryDelete);
-
-                if ($resultadoConsulta) {
-                    header('Location: /admin?resultado=3');
-                } else {
-                    echo "Error al eliminar la propiedad: " . mysqli_error($db);
-                }
-            } else {
-                echo "No se encontró la propiedad.";
             }
         } else {
             echo "ID no válido.";
@@ -84,10 +115,16 @@ if (!$resultadoConsulta) {
         <p class="alerta exito">Anuncio actualizado correctamente</p>
     <?php elseif ($resultado == '3'): ?>
         <p class="alerta error">Anuncio Borrado correctamente</p>
+    <?php elseif ($resultado == '4'): ?>
+        <p class="alerta error">Vendedor Borrado correctamente</p>
+    <?php elseif ($resultado == '5'): ?>
+        <p class="alerta error">Vendedor Actualizado correctamente</p>
     <?php endif ?>
 
     <a href="/admin/propiedades/crear.php" class="boton boton-verde"> Nueva Propiedad</a>
+    <a href="/admin/vendedores/crear.php" class="boton boton-amarillo"> Nuevo(a) vendedor</a>
 
+    <h2>Propiedades</h2>
     <table class="propiedades">
         <thead>
             <tr>
@@ -101,7 +138,7 @@ if (!$resultadoConsulta) {
 
         <tbody><!--Mostrar los resultados-->
             <!-- recorrer los resultados -->
-            <?php foreach( $propiedades as $propiedad ): ?>
+            <?php foreach ($propiedades as $propiedad): ?>
                 <tr>
                     <td><?php echo $propiedad->id; ?></td>
                     <td><?php echo $propiedad->titulo; ?></td>
@@ -110,6 +147,7 @@ if (!$resultadoConsulta) {
                     <td>
                         <form action="" method="POST" class="w-100">
                             <input type="hidden" name="id" value="<?= $propiedad->id; ?>">
+                            <input type="hidden" name="tipo" value="propiedad">
                             <input type="submit" class="boton-rojo-block" value="Eliminar">
                         </form>
                         <a href="propiedades/actualizar.php?id=<?php echo $propiedad->id; ?>" class="boton-amarillo-block">Actualizar</a>
@@ -118,6 +156,38 @@ if (!$resultadoConsulta) {
             <?php endforeach; ?>
         </tbody>
     </table>
+    <h2>Vendedores</h2>
+
+    <table class="propiedades">
+        <thead>
+            <tr>
+                <th>ID</th>
+                <th>Nombre</th>
+                <th>Acciones</th>
+            </tr>
+        </thead>
+
+        <tbody><!--Mostrar los resultados-->
+            <?php foreach ($vendedores as $vendedor): ?>
+                <tr>
+                    <td><?php echo $vendedor->vendedor_id; ?></td>
+                    <!-- quiero mostrar el nombre y apellido del vendedor de una sola vez
+                      -->
+                    <td><?php echo $vendedor->nombre . " " . $vendedor->apellido; ?></td>
+                    
+                    <td> <!-- Acciones -->
+                        <form action="" method="POST" class="w-100">
+                            <input type="hidden" name="id" value="<?= $vendedor->vendedor_id; ?>">
+                            <input type="hidden" name="tipo" value="vendedor">
+                            <input type="submit" class="boton-rojo-block" value="Eliminar">
+                        </form>
+                        <a href="vendedores/actualizar.php?id=<?php echo $vendedor->vendedor_id; ?>" class="boton-amarillo-block">Actualizar</a>
+                    </td>
+                </tr>
+            <?php endforeach; ?>
+        </tbody>
+    </table>
+
 </main>
 
 <footer class="footer seccion">
